Minor edits in types

+++ app/src/main/java/com/example/resistorcalculation/AllHistory.kt
@@ -5,7 +5,7 @@ class AllHistory {
     var voltage: Double = 0.0
     var f_voltage: Double = 0.0
     var current: Double = 0.0
-    var n_o_leds: Double = 0.0
+    var n_o_leds: Int = 0
     var resistence: Double = 0.0
     var min_power: Double = 0.0
 }
